+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 * Fix HTML injection in search form in the resource overview widget
 * Add a new text style to select in the rich text editor.
 * Allow counter-widgets to display the argument count for a specific idea
+* Add options to hide votes / status / last name in resource-representation-widgets
+* Allow title above share buttons to be changed in resource-representation-widgets
 
 ## v0.20.2
 * Fix resource overview widget error: TypeError: Cannot read property 'automaticallyUpdateStatus' of undefined